+++ ai-concept-demo/concept-demo.py
@@ -5,41 +5,13 @@ import strictyaml
 import networkx as nx
 import matplotlib.pyplot as plt
 from matplotlib import font_manager, rc
+import sys, os
+sys.path.append(os.path.dirname(os.path.abspath(os.path.dirname(__file__))))
+from coursegraph import fontutil as font #폰트 모듈 호출
+
 
 script_path = os.path.abspath(__file__)
 script_dir = os.path.dirname(script_path)
-
-# OS 확인 후 폰트 선택 하도록 수정
-#맥OS에서 글자 깨짐 현상 해결
-def get_system_font():
-    flist = font_manager.findSystemFonts()
-    data_list = []
-    system = platform.system()
-
-    if system == 'Windows':
-        # Windows의 경우, 시스템 폰트 경로로 설정
-        for v in flist:
-            try:
-                fprop = font_manager.FontProperties(fname=v)
-                if fprop.get_name() == 'Malgun':
-                    data_list.append({"name": fprop.get_name(), "file": fprop.get_file()})
-            except:
-                continue
-        return [data_list[0]['file']]
-    elif system == 'Darwin':
-        for v in flist:
-            try:
-                fprop = font_manager.FontProperties(fname=v)
-                if fprop.get_name() == 'Apple SD Gothic Neo':
-                    data_list.append({"name": fprop.get_name(), "file": fprop.get_file()})
-            except:
-                continue
-        # macOS의 경우, 기본 시스템 폰트 경로 설정 (MacOS 애플고딕의 정확한 이름 표기가 필요함)
-        return [data_list[0]['file']]
-    else:
-        # 기타 운영 체제의 경우, 적절한 시스템 폰트 경로를 설정해야 합니다. (현재 리눅스의 파일 경로에 맞춰 설정되어있음) ===> 리눅스에 맞지 않음!! 배포판마다 위치 다름 이런 하드코딩 정말 안좋습니다
-        # return ['/usr/share/fonts/truetype/NanumGothic.ttf'] 
-        return None
 
 def read_subjects(filename):
     try:
@@ -75,9 +47,11 @@ def adjust_coordinates(subjects):
     return adjusted_pos
 
 def draw_course_structure(subjects):
-    system_fonts = get_system_font()
+    fonts = font.get_system_font()
+    plt.rcParams['font.family'] = fonts[0]['file']
+    system_fonts = fonts[0]['file']
     if system_fonts:
-        font_path = system_fonts[0]
+        font_path = system_fonts
         font_name = font_manager.FontProperties(fname=font_path).get_name()
     else:
         # 무조건 NanumGothic 시도하도록 임시로
